move to lower_to_communication
